requirements.txt
fastapi[standard]==0.113.0
sqlalchemy
pydantic-settings
python-multipart
